+++ x/perp/keeper/perp_test.go
@@ -176,6 +176,7 @@ func TestKeeperClosePosition(t *testing.T) {
 		)
 
 		t.Log("open position for alice - long")
+		ctx = ctx.WithBlockHeight(ctx.BlockHeight() + 1).WithBlockTime(time.Now().Add(time.Minute))
 
 		alice := sample.AccAddress()
 		err := simapp.FundAccount(nibiruApp.BankKeeper, ctx, alice,
@@ -206,6 +207,7 @@ func TestKeeperClosePosition(t *testing.T) {
 		bobQuote := sdk.NewInt(60)
 		bobLeverage := sdk.NewDec(10)
 		bobBaseLimit := sdk.NewDec(150)
+
 		_, err = nibiruApp.PerpKeeper.OpenPosition(
 			ctx, pair, bobSide, bob, bobQuote, bobLeverage, bobBaseLimit)
 		require.NoError(t, err)