+++ x/perp/keeper/clearing_house_unit_test.go
@@ -886,6 +886,7 @@ func TestClosePositionEntirely(t *testing.T) {
 				ctx,
 				tc.initialPosition,
 				/*quoteAssetLimit=*/ tc.quoteAssetLimit, // NUSD
+				/* skipFluctuationLimitCheck */ false,
 			)
 
 			require.NoError(t, err)