+++ scripts/localnet.sh
@@ -61,7 +61,7 @@ fi
 
 # Initialize nibid with "localnet" chain id
 echo_info "Initializing $CHAIN_ID..."
-if $BINARY init nibiru-localnet-0 --home $CHAIN_HOME_DIR --chain-id $CHAIN_ID; then
+if $BINARY init nibiru-localnet-0 --home $CHAIN_HOME_DIR --chain-id $CHAIN_ID --overwrite; then
   echo_success "Successfully initialized $CHAIN_ID"
 else
   echo_error "Failed to initialize $CHAIN_ID"