+++ CHANGELOG.md
@@ -42,6 +42,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Improvements
+
+* [#775](https://github.com/NibiruChain/nibiru/pull/775) - bump google.golang.org/protobuf from 1.28.0 to 1.28.1
+
 ### Bug Fixes
 
 * [#766](https://github.com/NibiruChain/nibiru/pull/766) Fixed margin ratio calculation for trader position.