+++ CHANGELOG.md
@@ -45,7 +45,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Improvements
 
 * [#775](https://github.com/NibiruChain/nibiru/pull/775) - bump google.golang.org/protobuf from 1.28.0 to 1.28.1
-* [#768](https://github.com/NibiruChain/nibiru/pull/768/) - add simulation tests to make file
+* [#768](https://github.com/NibiruChain/nibiru/pull/768) - add simulation tests to make file
+* [#767](https://github.com/NibiruChain/nibiru/pull/767) - add fluctuation limit checks on `OpenPosition`.
 
 ### Bug Fixes
 
@@ -125,4 +126,3 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 * [#695](https://github.com/NibiruChain/nibiru/pull/695) Add `OpenPosition` integration tests.
 * [#692](https://github.com/NibiruChain/nibiru/pull/692) Add test coverage for Perp MsgServer methods.
-* [#768](https://github.com/NibiruChain/nibiru/pull/768) Update simulation tests to read flags over static values