+++ CHANGELOG.md
@@ -52,6 +52,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 * [#766](https://github.com/NibiruChain/nibiru/pull/766) Fixed margin ratio calculation for trader position.
 
+### Testing
+* [#782](https://github.com/NibiruChain/nibiru/pull/782) - replace github test workflows to use make commands
+
 ## [v0.11.0](https://github.com/NibiruChain/nibiru/releases/tag/v0.11.0) - 2022-07-29
 
 ### Documentation